C: use long's for comparison.
- Long'er is better.

+++ c/unit-test.h
@@ -48,7 +48,7 @@ void _expectString(const char *, const char *, int);
 */
 #define expectString(s1, s2) _expectString(s1, s2, __LINE__)
 
-void _expectEquals(int val1, int val2, int line);
+void _expectEquals(long val1, long val2, int line);
 /** Standard expect of values.
 
     @param val1, The first value.
